AMD IOMMU: bugref:9654 IOTLB cache bits.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@87327 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Devices/Bus/DevIommuAmd.cpp
@@ -134,7 +134,7 @@ typedef struct IOWALKRESULT
     /** The I/O permissions allowed by the translation (IOMMU_IO_PERM_XXX). */
     uint8_t         fIoPerm;
     /** Padding. */
-    uint16_t        abPadding[2];
+    uint8_t         abPadding[2];
 } IOWALKRESULT;
 /** Pointer to an I/O walk result struct. */
 typedef IOWALKRESULT *PIOWALKRESULT;
